// src/apis/qrcode/type.ts
export type qrCodeModel = {
  owner?: {
    _id: string;
    userName: string;
    fullName: string;
    profileImage: string;
  };
  webSite?: string;
  companyProfile?: string;
  whatsApp?: string;
  faceBook?: string;
  linkedIn?: string;
  instagram?: string;
  threads?: string;
  snapChat?: string;
  youtube?: string;
  tiktok?: string;
  painterest?: string;
  xPlatform?: string;
  otherLink?: string;
  location?: string;
};